+++ packages/backend/src/nest/libp2p/libp2p.const.ts
@@ -1,5 +1,6 @@
 export const BITSWAP_PROTOCOL = '/quiet/ipfs/bitswap/1.2.0'
 export const AUTH_STREAM_TIMEOUT_MS = 10_000
+export const PEER_RETRY_DELAY_MS = 8_000
 export const WEBSOCKET_CIPHER_SUITE = 'TLS_AES_256_GCM_SHA384'
 
 export const UNKNOWN_REMOTE_PEER = `The peer's device isn't listed on this team`